requirements.txt
google-adk>=0.2.0
google-generativeai>=0.8.0
pypdf>=4.0.0
python-dotenv>=1.0.0